created separate stylesheet
